==> docs/01_prep_load.Rmd <==
---
title: "01 Load IISS Data"
author: "Andres Gannon"
date: "April 6, 2019"
output:
  html_document:
    theme: flatly
    code_download: true
    toc: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---
<style>
    body .main-container {
        max-width: 100%;
    }
</style>

```{r setup, include=FALSE}
rm(list=ls())
knitr::opts_chunk$set(echo = TRUE)
```

This document loads and then saves the IISS equipment data

# Set up
```{r}
library(magrittr)
library(tidyverse)
```

# Manual load
r import merge excel files multiple sheets site:stackoverflow.com
```{r}
files <- list.files(paste0(here::here(), '/data/original/'), "\\.xlsx$", full.names = TRUE)

read_excel_allsheets <- function(filename) { 
  sheets <- readxl::excel_sheets(filename) 
  x <- lapply(sheets, function(X) readxl::read_excel(filename, sheet = X)) 
  names(x) <- sheets 
  x 
} 

tictoc::tic()
out <- lapply(files, read_excel_allsheets)
tictoc::toc()

basename(files)
```

# Join sheets for each year  
```{r}
# Initiate empty list
by_year = list()

# By removing the assign function, you can avoid printing all the datasets into the workspace
for(i in 1:54){
  by_year[i] = list(assign(paste0('IISS_equipdata_', i+1960), plyr::ldply(out[[i]], data.frame)))
}
```

# Join all data and save file
```{r}
IISS_full <- do.call(rbind.data.frame, by_year)

# Remove .id column 
IISS_full[,".id"] <- NULL

saveRDS(IISS_full, paste0(here::here(), "/data/","01a_load.rds"))
```

# System info
```{r}
sessionInfo()
```
