==> docs/02_prep_clean.Rmd <==
---
title: "01 Prep IISS Data"
author: "J Andres Gannon, Erin Werner, and Tom Brailey"
date: "03/24/18"
output:
  html_notebook:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  html_document:
    toc: yes
editor_options:
  chunk_output_type: inline
---
<style>
    body .main-container {
        max-width: 100%;
    }
</style>


This is the entry point for the IISS dataset.

In this file, a raw csv file of the data on national military platforms is loaded and processed.

# Load Library
```{r , results='hide', message=FALSE, warning=FALSE, include=FALSE}
# Reset workspace
rm(list = ls())

# Memory saver
gc()

# Load essential packages
library(magrittr)
library(ggplot2)
library(data.table)
library(DescTools)

# Source functions
source(paste0(here::here(), "/R/missingness_df.R"))
source(paste0(here::here(), "/R/parentchild.R"))
source(paste0(here::here(), "/R/equip_domain.R"))
source(paste0(here::here(), "/R/country_year_equip_df.R"))
source(paste0(here::here(), "/R/iiss_wrangling_function.R"))

# Aesthetics
knitr::opts_knit$set(progress = TRUE, verbose = TRUE)
knitr::opts_chunk$set(fig.width = 12, fig.height = 8, warning = FALSE, message = FALSE, cache = TRUE)
options(width = 160, knitr.duplicate.label = 'allow')
```


# Load raw data for cleaning and diagnostics
All data is entered in google docs, but then manually downloaded as excel files that are combined into a .rds

00_LoadData.Rmd converts the by-country .xlsx files into one .rds file.
```{r}
# Load .rds file
iiss <- readRDS(file = paste0(here::here(), '/data/01a_load.rds')) %>%
  janitor::remove_empty("rows") %>%
  dplyr::mutate(across(where(is.character), stringr::str_trim))
```

# Add country name variables
To make merging with other datasets easier, we need to include other variations of country names.
```{r include = FALSE}
country_meta <- read.csv(file = paste0(here::here(), '/data/aux/IISS_supp_iiss_countries.csv'))
```

# Column checks
## Year
Checks for empty year rows
```{r}
iiss$year <- as.numeric(iiss$year)

sum(is.na(iiss$year))
missing_years <- dplyr::filter(iiss, is.na(year)) %>%
  dplyr::select(-c("service", "subservice", "subsubservice"))

sort(unique(missing_years$country))
```

## Country
Fixes manual coding errors in the country column
```{r}
iiss$country <- as.factor(iiss$country)

iiss$subservice[iiss$country == "pre-strategic nuclear forces"] <- "pre-strategic nuclear forces"
iiss$country[iiss$country == "pre-strategic nuclear forces"] <- "france"
iiss$country[iiss$country == "Antigua and Barbuda"] <- "antigua and barbuda"
iiss$country[iiss$country == "Argentina"] <- "argentina"
iiss$country[iiss$country == "palestinian autonomous areas of Gaza and Jericho"] <- "palestinian territories"
iiss$country[iiss$country == "Iran"] <- "iran"
iiss$country[iiss$country == "Laos"] <- "laos"
iiss$country[iiss$country == "Jamaica"] <- "jamaica"
iiss$country[iiss$country == "mail"] <- "mali"

sum(is.na(iiss$country))
missing_countries <- dplyr::filter(iiss, is.na(country)) %>% 
  dplyr::select(-c("service", "subservice", "subsubservice"))

### Remove multinational organization
row_to_keep = which(iiss$country == "multinationalorganization")
iiss <<- iiss[-row_to_keep, ]

test <- iiss %>%
  dplyr::select(year, country)

sort(unique(missing_countries$year))
```

## Continent
Adds continent variables and manually adds country codes for states without values
```{r}
# Add continent data
# Serbia needs a manual input for its cown ccode.
iiss$continent <- countrycode::countrycode(iiss$country, 'country.name', 'continent')

# Only serbia does not have a code listed under cown
iiss$ccode <- countrycode::countrycode(iiss$country, 'country.name', 'cown')
iiss$ccode[iiss$country == "serbia"] <- "345"

# There may be other non-European MNOs in other years of the data
iiss$continent[iiss$country == "multinationalorganization"] <- "europe"
iiss$ccode[iiss$country == "multinationalorganization"] <- "000"
iiss$ccode[iiss$country == "palestinian territories"] <- "665"

# Both should be zero
sum(is.na(iiss$continent))
sum(is.na(iiss$ccode))
```

## Domain
New domain column based on values of service
```{r}
origS <- c("Joint Forces Command (JFC)","army","air and air defence aviation forces",
           "paramilitary","space","navy","air force","joint command",
           "air force and air defence","land component","naval component",
           "air component","armed forces","royal canadian navy",
           "royal canadian air force","NATO flight training canada",
           "contracted flying services - southport",
           "canadian special operations forces command","canadian coast guard",
           "royal canadian mounted police","reserves","strategic nuclear forces",
           "land","air","air corps","joint special operations command",
           "joint operational command","home guard","land forces command",
           "navy - maritime action force","land forces","strategic forces",
           "joint helicopter command","STRATCOM","Space","Army","Navy",
           "Marine Corps","Coast Guard","Air Force","Special Operations Command",
           "ground forces (army)","airborne forces",NA,"air and AD joint command",
           "air force and air defence forces","air wing","armed police",
           "maritime squadron","national guard","NATO aewc",
           "strategic airlift capability","strategic airlift interim solution",
           "national army","strategic missile forces","strategic deterrent forces",
           "space forces","military air forces","air force/air defence",
           "strategic forces command","coast guard","ground self-defense force",
           "maritime self-defense force","air self-defense force","naval aviation",
           "marines","royal bahamian defence force",
           "revolutionary guard corps","naval",
           "royal household","maritime","air defence forces","air defence command",
           "missile command","air defence","gendarmerie","air forces",
           "anti-aircraft defence and revolutionary air force",
           "people's defence force", "armed  forces",
           "department of agriculture, fisheries, and forestry",
           "department of environmental affairs","somaliland"
)

# Service typos

# for rows with missing service values
missing_service <- dplyr::filter(iiss, is.na(service)) %>% dplyr::distinct(country, year)

serviceB <- unique(iiss$service)
if(length(serviceB) > length(origS)){
  #need to add a new type of service to the list of services
  print("Update list of services with new type of service!")
  return
}

domain <- c("land", "air", "sea", "space", "cyber", "nuclear")

unique(iiss$service)

iiss_serviceview <- iiss %>%
  dplyr::select(year, country, service, subservice, equipment_type) %>%
  dplyr::distinct()

# (based on overall/average equipment types under that branch of service)
# Special Operations Command  and similar variants- air
# STRATCOM - cyber
iiss$domain <- ifelse(iiss$service=="Space" |
                        iiss$service=="space forces" |
                        iiss$service=="space",      "space",

                      ifelse(iiss$service=="army" | 
                               iiss$service=="Army" |
                               iiss$service=="paramilitary" |
                               iiss$service=="Joint Forces Command (JFC)" |
                               iiss$service=="joint command" |
                               iiss$service=="land component" | 
                               iiss$service=="armed forces" |
                               iiss$service=="canadian special operations forces command" |
                               iiss$service=="royal canadian mounted police" |
                               iiss$service=="reserves" | 
                               iiss$service=="land" |
                               iiss$service=="home guard" |
                               iiss$service=="land forces command" | 
                               iiss$service=="land forces" |
                               iiss$service=="Marine Corps" |
                               iiss$service=="marines" |
                               iiss$service=="armed police" |
                               iiss$service=="department of agriculture, fisheries, and forestry" |
                               iiss$service=="national guard" |
                               iiss$service=="revolutionary guard corps" |
                               iiss$service=="royal bahamian defence force" |
                               iiss$service=="national army" |
                               iiss$service=="ground self-defense force" |
                               iiss$service=="royal household" |
                               iiss$service=="gendarmerie" |
                               iiss$service=="armed  forces" |
                               iiss$service=="somaliland" |
                               iiss$service=="department of environmental affairs" |
                               iiss$service=="people's defence force" |
                               iiss$service=="strategic deterrent forces" |
                               iiss$service=="ground forces (army)",           "land",

                             ifelse(iiss$service=="navy" | 
                                      iiss$service =="Navy" |
                                      iiss$service=="naval component" |
                                      iiss$service=="royal canadian navy" |
                                      iiss$service=="canadian coast guard" |
                                      iiss$service=="maritime squadron" |
                                      iiss$service=="coast guard" |
                                      iiss$service=="naval" |
                                      iiss$service=="maritime self-defense force" |
                                      iiss$service=="naval aviation" |
                                      iiss$service=="maritime" |
                                      iiss$service=="navy - maritime action force" |
                                      iiss$service=="Coast Guard",                    "sea",

                                    ifelse(iiss$service=="air and air defence aviation forces" |
                                             iiss$service=="air force" |
                                             iiss$service=="air force and air defence" |
                                             iiss$service=="air component" |
                                             iiss$service=="royal canadian air force" |
                                             iiss$service=="NATO flight training canada" |
                                             iiss$service=="contracted flying services - southport" |
                                             iiss$service=="air" | 
                                             iiss$service=="air corps" |
                                             iiss$service=="joint helicopter command" |
                                             iiss$service=="Special Operations Command" |
                                             iiss$service=="joint special operations command" |
                                             iiss$service=="joint operational command" |
                                             iiss$service=="air and AD joint command" |
                                             iiss$service=="air wing" |
                                             iiss$service=="NATO aewc" |
                                             iiss$service=="air forces" |
                                             iiss$service=="air self-defense force" |
                                             iiss$service=="strategic missile forces" |
                                             iiss$service=="military air forces" |
                                             iiss$service=="air force/air defence" |
                                             iiss$service=="strategic airlift interim solution" |
                                             iiss$service=="strategic airlift capability" |
                                             iiss$service=="air force and air defence forces" |
                                             iiss$service=="air defence forces" |
                                             iiss$service=="air defence command" |
                                             iiss$service=="missile command" |
                                             iiss$service=="anti-aircraft defence and revolutionary air force" |
                                             iiss$service=="air defence" |
                                             iiss$service=="Air Force" | 
                                             iiss$service=="airborne forces",   "air",

                                           ifelse(iiss$service=="STRATCOM","cyber",

                                                  ifelse(iiss$service=="strategic nuclear forces" |
                                                           iiss$service=="strategic forces command" |
                                                           iiss$service=="strategic forces",                 "nuclear",
                                                         NA  )))))) # all other values map to NA
```

## Service Columns
Fixes manual coding errors in the services columns

### Service
```{r}
# Type in coastal reserve
iiss$service[iiss$service == "costal defence"] <- "coastal defence"

# remove irrelevant services 
row_to_keep = which(iiss$service == "foreign forces" |
                      iiss$service == "forces abroad")
iiss <<- iiss[-row_to_keep, ]

# to lower
iiss$service <- tolower(iiss$service)
```

### Subservice
```{r}
# coerce reserve
iiss$subservice[iiss$subservice == "reserve"] <- "reserves"
```

### Subsubservice
```{r}
# No issues yet
```

## Unit Count
Cleans unit_count strings
```{r}
# updates notes then converts to int
iiss$notes[iiss$unit_count %like% "%\\+%"] <- "minimum value"

iiss$unit_count[iiss$equipment_type == "bombs"] <- 0
iiss$unit_count[iiss$unit_count=="some"] <- NA

# updates notes then converts to int
iiss$unit_count <- gsub('+', '', iiss$unit_count)
iiss$unit_count <- gsub('*', '', iiss$unit_count)

iiss$unit_count <- as.numeric(gsub("[^0-9\\.]", "", iiss$unit_count))
```

## Raw Text
Addresses any issues with the rawtext columns
```{r}
iiss$rawtext[iiss$country == "timor leste" &
               iiss$rawtext == "NA"] <- "No equipment information given"
```

## New/Old Row
Makes sure old/new_row's are all zeros and ones
```{r}
iiss$new_row[iiss$new_row == "`"] <- NA
iiss$new_row[iiss$new_row == "no counts given"] <- NA
iiss$new_row[iiss$new_row == "21"] <- 1
iiss$new_row[iiss$new_row == "submarines and Thetis have capability"] <- NA
iiss$new_row[iiss$new_row == "174 in store"] <- NA
iiss$new_row <- as.numeric(iiss$new_row)
unique(iiss$new_row)

iiss$old_row[iiss$old_row == "`"] <- NA
iiss$old_row[iiss$old_row == "no counts given"] <- NA
iiss$old_row[iiss$old_row == "11"] <- 1
iiss$old_row[iiss$old_row == "2"] <- 1
iiss$old_row <- as.numeric(iiss$old_row)
unique(iiss$old_row)

old_check <- iiss[which(iiss$old_row==1),]
zero_check <- nrow(old_check[which(old_check$unit_count==0 | is.na(old_check$unit_count)),])
percentage <- zero_check/nrow(old_check)
# should be 100%
percentage*100

iiss$new_row[iiss$year==2017 & iiss$country=="us"] <- NA
iiss$old_row[iiss$year==2017 & iiss$country=="us"] <- NA

iiss$new_row[iiss$year==2014 & iiss$country!="us"] <- NA
iiss$old_row[iiss$year==2014 & iiss$country!="us"] <- NA

# if old_row = 1, then unit_count = 0. If not, return row
c <- nrow(old_check[which(old_check$unit_count>0),])
if(c>0){
  d <- old_check[which(old_check$unit_count>0),]
  rownames(d)
}

new_check <- iiss[which(iiss$new_row==1),]
zero_check <- nrow(new_check[which(new_check$unit_count>0 | is.na(new_check$unit_count)),])
percentage <- zero_check/nrow(new_check)
# should be 100%
percentage*100

# if unit_count > 0, old_row = blank
iiss$old_row[iiss$unit_count > 0] <- NA

# if unit_count = n/a | NA | N/A | some and new_row = 1, unit_count = blank
iiss$unit_count[is.na(iiss$unit_count) & iiss$new_row==1] <- NA

# if unit_count = 0 and new_row = 1, do nothing
# if unit_count = blank and new_row = 1, do nothing
```

## Notes
Addresses coder notes and updates notes section when issues are fixed
```{r}
iiss$notes[iiss$notes == "\"some\""] <- "some"
iiss$notes[iiss$notes == "In storage"] <- "in storage"
iiss$notes[iiss$notes == "in store"] <- "in storage"
iiss$notes[iiss$notes == "many in store"] <- "in storage"
iiss$notes[iiss$notes == "mean value used"] <- "mean value taken"
iiss$notes[iiss$notes == "minimum value"] <- "minimum"
iiss$notes[iiss$notes == "minmum"] <- "minimum"
iiss$notes[iiss$notes == "no number given"] <- "no count given"
iiss$notes[iiss$notes == "no numbers given"] <- "no count given"
iiss$notes[iiss$notes == "no given number"] <- "no count given"
iiss$notes[iiss$notes == "testing"] <- "in test"
iiss$notes[iiss$notes == "not sure about this equipment type, there is no top level header provided"] <- NA
iiss$notes[iiss$notes == "not sure about this equipment typeand there is no top level header provided"] <- NA
iiss$notes[iiss$notes == "plus up to 300 in store"] <- NA
iiss$notes[iiss$notes == "probably out of service"] <- NA
iiss$notes[iiss$notes %like% "%\\+%"] <- "minimum"
iiss$notes[iiss$notes %like% "%up to%"] <- "maximum"
iiss$notes[iiss$notes == "check unit name, looks odd"] <- NA
iiss$notes[iiss$notes == "check unit nameand looks odd"] <- NA
iiss$notes[iiss$notes == "1.0"] <- NA

#sort(unique(iiss$notes))

unique_notes <- data.frame(dplyr::distinct(iiss, notes, .keep_all = TRUE))
```

# Wrangling
## Function code

The file 01b_prepIISS_supp.Rmd deals with typos and formatting errors associated with military equipment (equipment_type, equipment_subtype, equipment_name, equipment_subname, unit_name). The file is read in and run two times to ensure that all coding errors are fixed. 

```{r}
# Create function to read in .Rmd files
source_rmd <- function(file, local = FALSE, ...){
  options(knitr.duplicate.label = 'allow')

  tempR <- tempfile(tmpdir = ".", fileext = ".R")
  on.exit(unlink(tempR))
  knitr::purl(file, output=tempR, quiet = TRUE)

  envir <- globalenv()
  source(tempR, local = envir, ...)
}

# Run the function 2 times 
tictoc::tic()

for(i in c(1:2)){
  source_rmd(paste0(here::here(), "/docs/02aux_prep_clean-supp.Rmd"))
}

tictoc::toc()
```

## Check each equipment_type

Having run the wrangling code, we can better understand the relationship between equipment types, subtypes, names, and subnames for each distinct equipment_type in the dataset. We use these subsetted data to make note of any remaining errors, and then add the relevant cleaning code to 01b_prepIISS_supp.Rmd. 

*2022-04-26 Fix the following equipment types:*
miscellaneous = UK 1983-1985 repair forwardrepair

### Unique equipment_types
```{r}
sort(unique(iiss$equipment_type))

# write.csv(iiss, paste0(here::here(), "/data/","iiss_full.csv"))
```

### Miscellaneous
```{r}
# Checks for empty equipment_type rows
iiss %>%
  dplyr::filter(is.na(equipment_type)) %>% 
  dplyr::distinct(country, year, equipment_subtype, equipment_name, unit_name) %>%
  DT::datatable()

# Check for empty equipment_name rows
iiss %>%
  dplyr::select(equipment_type, equipment_name, unit_name) %>%
  dplyr::filter(is.na(equipment_name)) %>%
  dplyr::distinct() %>%
  dplyr::select(-equipment_name) %>%
  dplyr::arrange(equipment_type, unit_name) %>%
#  write.csv(paste0(here::here(), "/R/cleaning/missing_equipmentname.csv")) %>%
  DT::datatable()

# Check for duplicate rows
iiss_dup <- iiss %>%
  dplyr::select(country, year, service, subservice, equipment_type, equipment_subtype, equipment_name, equipment_subname, unit_name, unit_count) %>%
  dplyr::group_by(country, year, service, subservice, equipment_type, equipment_subtype, equipment_name, equipment_subname, unit_name, unit_count) %>%
  dplyr::filter(!is.na(unit_count), 
                unit_count != 0,
                n() > 1)
write.csv(iiss_dup, paste0(here::here(), "/R/cleaning/potential_duplicates.csv"))

# Look at percentage of non-zero but unspecified quantities of equipments by equipment_name
iiss %>%
  dplyr::group_by(equipment_type, equipment_name) %>%
  dplyr::summarize(perc_na = round(sum(is.na(unit_count)) / length(unit_name), 2) * 100) %>%
  DT::datatable()

iiss %>%
  dplyr::group_by(equipment_type) %>%
  dplyr::summarize(perc_na = round(sum(is.na(unit_count)) / length(unit_name), 2) * 100) %>%
  DT::datatable()
```

### air defence
```{r}
ad_checks <- iiss %>% 
  dplyr::filter(equipment_type == "air defence")
sort(unique(ad_checks$equipment_subtype))
sort(unique(ad_checks$equipment_name))
sort(unique(ad_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(ad_checks$equipment_type))
est <- sort(unique(ad_checks$equipment_subtype))
en <- sort(unique(ad_checks$equipment_name))
esn <- sort(unique(ad_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "air defence") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

ad_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 600, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### air-launched missiles
```{r}
alm_checks <- iiss %>% 
  dplyr::filter(equipment_type == "air-launched missiles")
sort(unique(alm_checks$equipment_subtype))
sort(unique(alm_checks$equipment_name))
sort(unique(alm_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(alm_checks$equipment_type))
est <- sort(unique(alm_checks$equipment_subtype))
en <- sort(unique(alm_checks$equipment_name))
esn <- sort(unique(alm_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "air-launched missiles") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

alm_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 600, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### aircraft
```{r}
aircraft_checks <- iiss %>% 
  dplyr::filter(equipment_type == "aircraft")
sort(unique(aircraft_checks$equipment_subtype))
sort(unique(aircraft_checks$equipment_name))
sort(unique(aircraft_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(aircraft_checks$equipment_type))
est <- sort(unique(aircraft_checks$equipment_subtype))
en <- sort(unique(aircraft_checks$equipment_name))
esn <- sort(unique(aircraft_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "aircraft") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

aircraft_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 6000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### amphibious
```{r}
amphibious_checks <- iiss %>% 
  dplyr::filter(equipment_type == "amphibious") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(amphibious_checks$equipment_subtype))
sort(unique(amphibious_checks$equipment_name))
sort(unique(amphibious_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(amphibious_checks$equipment_type))
est <- sort(unique(amphibious_checks$equipment_subtype))
en <- sort(unique(amphibious_checks$equipment_name))
esn <- sort(unique(amphibious_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "amphibious") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

amphibious_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 6000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### anti-tank/anti-infrastructure
```{r}
atai_checks <- iiss %>% 
  dplyr::filter(equipment_type == "anti-tank/anti-infrastructure") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(atai_checks$equipment_subtype))
sort(unique(atai_checks$equipment_name))
sort(unique(atai_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(atai_checks$equipment_type))
est <- sort(unique(atai_checks$equipment_subtype))
en <- sort(unique(atai_checks$equipment_name))
esn <- sort(unique(atai_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "anti-tank/anti-infrastructure") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

atai_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 6000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### armoured fighting vehicles
```{r}
afv_checks <- iiss %>% 
  dplyr::filter(equipment_type == "armoured fighting vehicles") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(afv_checks$equipment_subtype))
sort(unique(afv_checks$equipment_name))
sort(unique(afv_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(afv_checks$equipment_type))
est <- sort(unique(afv_checks$equipment_subtype))
en <- sort(unique(afv_checks$equipment_name))
esn <- sort(unique(afv_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "armoured fighting vehicles") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

afv_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 6000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### artillery
```{r}
artillery_checks <- iiss %>% 
  dplyr::filter(equipment_type == "artillery") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(artillery_checks$equipment_subtype))
sort(unique(artillery_checks$equipment_name))
sort(unique(artillery_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(artillery_checks$equipment_type))
est <- sort(unique(artillery_checks$equipment_subtype))
en <- sort(unique(artillery_checks$equipment_name))
esn <- sort(unique(artillery_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "artillery") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

artillery_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 6000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### ballistic missiles
```{r}
bm_checks <- iiss %>% 
  dplyr::filter(equipment_type == "ballistic missiles") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(bm_checks$equipment_subtype))
sort(unique(bm_checks$equipment_name))
sort(unique(bm_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(bm_checks$equipment_type))
est <- sort(unique(bm_checks$equipment_subtype))
en <- sort(unique(bm_checks$equipment_name))
esn <- sort(unique(bm_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "ballistic missiles") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

bm_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 800, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### bombs
```{r}
bombs_checks <- iiss %>% 
  dplyr::filter(equipment_type == "bombs") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(bombs_checks$equipment_subtype))
sort(unique(bombs_checks$equipment_name))
sort(unique(bombs_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(bombs_checks$equipment_type))
est <- sort(unique(bombs_checks$equipment_subtype))
en <- sort(unique(bombs_checks$equipment_name))
esn <- sort(unique(bombs_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "bombs") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

bombs_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 800, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### engineering and maintenance vehicles
```{r}
engineering_checks <- iiss %>% 
  dplyr::filter(equipment_type == "engineering and maintenance vehicles") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(engineering_checks$equipment_subtype))
sort(unique(engineering_checks$equipment_name))
sort(unique(engineering_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(engineering_checks$equipment_type))
est <- sort(unique(engineering_checks$equipment_subtype))
en <- sort(unique(engineering_checks$equipment_name))
esn <- sort(unique(engineering_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "engineering and maintenance vehicles") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

engineering_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 400, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### helicopters
```{r}
helicopters_checks <- iiss %>% 
  dplyr::filter(equipment_type == "helicopters") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(helicopters_checks$equipment_subtype))
sort(unique(helicopters_checks$equipment_name))
sort(unique(helicopters_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(helicopters_checks$equipment_type))
est <- sort(unique(helicopters_checks$equipment_subtype))
en <- sort(unique(helicopters_checks$equipment_name))
esn <- sort(unique(helicopters_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "helicopters") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

helicopters_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 4000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### logistics and support
```{r}
logistics_checks <- iiss %>% 
  dplyr::filter(equipment_type == "logistics and support") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(logistics_checks$equipment_subtype))
sort(unique(logistics_checks$equipment_name))
sort(unique(logistics_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(logistics_checks$equipment_type))
est <- sort(unique(logistics_checks$equipment_subtype))
en <- sort(unique(logistics_checks$equipment_name))
esn <- sort(unique(logistics_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "logistics and support") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

logistics_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 4000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### mine warfare
```{r}
mines_checks <- iiss %>% 
  dplyr::filter(equipment_type == "mine warfare") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(mines_checks$equipment_subtype))
sort(unique(mines_checks$equipment_name))
sort(unique(mines_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(mines_checks$equipment_type))
est <- sort(unique(mines_checks$equipment_subtype))
en <- sort(unique(mines_checks$equipment_name))
esn <- sort(unique(mines_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "mine warfare") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

mines_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 1000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### patrol and coastal combatants
```{r}
pacc_checks <- iiss %>% 
  dplyr::filter(equipment_type == "patrol and coastal combatants") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(pacc_checks$equipment_subtype))
sort(unique(pacc_checks$equipment_name))
sort(unique(pacc_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(pacc_checks$equipment_type))
est <- sort(unique(pacc_checks$equipment_subtype))
en <- sort(unique(pacc_checks$equipment_name))
esn <- sort(unique(pacc_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "patrol and coastal combatants") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

pacc_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 1000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### principal surface combatants
```{r}
psc_checks <- iiss %>% 
  dplyr::filter(equipment_type == "principal surface combatants") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(psc_checks$equipment_subtype))
sort(unique(psc_checks$equipment_name))
sort(unique(psc_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(psc_checks$equipment_type))
est <- sort(unique(psc_checks$equipment_subtype))
en <- sort(unique(psc_checks$equipment_name))
esn <- sort(unique(psc_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "principal surface combatants") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

psc_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 1000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### radars
```{r}
radars_checks <- iiss %>% 
  dplyr::filter(equipment_type == "radars") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()

write.csv(radars_checks, file = paste0(here::here(), "/R/cleaning/radars_final_coerce.csv"))

sort(unique(radars_checks$equipment_subtype))
sort(unique(radars_checks$equipment_name))
sort(unique(radars_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(radars_checks$equipment_type))
est <- sort(unique(radars_checks$equipment_subtype))
en <- sort(unique(radars_checks$equipment_name))
esn <- sort(unique(radars_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "radars") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

radars_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 1000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```

### submarines
```{r}
submarines_checks <- iiss %>% dplyr::filter(equipment_type == "submarines")
sort(unique(submarines_checks$equipment_subtype))
sort(unique(submarines_checks$equipment_name))
sort(unique(submarines_checks$equipment_subname))
```

### surface-to-surface missile launchers
```{r}
stsml_checks <- iiss %>% 
  dplyr::filter(equipment_type == "surface-to-surface missile launchers") %>%
  dplyr::select(dplyr::starts_with("equipment_"), unit_name) %>%
  dplyr::distinct()
sort(unique(stsml_checks$equipment_subtype))
sort(unique(stsml_checks$equipment_name))
sort(unique(stsml_checks$equipment_subname))

# Values in more than one column
et <- sort(unique(stsml_checks$equipment_type))
est <- sort(unique(stsml_checks$equipment_subtype))
en <- sort(unique(stsml_checks$equipment_name))
esn <- sort(unique(stsml_checks$equipment_subname))

uniques <- as.data.frame(c(et, est, en, esn))
duplicated(uniques)

uniques <- iiss %>% 
  dplyr::filter(equipment_type == "surface-to-surface missile launchers") %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  dplyr::select(-equipment_type) %>%
  dplyr::distinct() %>%
  dplyr::arrange(equipment_subtype, equipment_name, equipment_subname)

stsml_checks %>%
  dplyr::select(dplyr::starts_with("equipment_")) %>%
  tidyr::replace_na(list(equipment_subtype = "NA_subtype",
                         equipment_name = "NA_name",
                         equipment_subname = "NA_subname")) %>%
  dplyr::distinct() %>%
  collapsibleTree::collapsibleTree(hierarchy = c("equipment_type",
                                                 "equipment_subtype",
                                                 "equipment_name",
                                                 "equipment_subname"),
                                   height = 1000, width = 800,
                                   collapse = FALSE, zoomable = FALSE)
```
### uav
```{r}
uav_checks <- iiss %>% dplyr::filter(equipment_type == "unmanned aerial vehicles")
sort(unique(uav_checks$equipment_subtype))
sort(unique(uav_checks$equipment_name))
sort(unique(uav_checks$equipment_subname))
```

### NA

There are several rows with "NA" in the equipment_type column. Where we have NA for country, year, and the equipment columns, we remove the entry. 
```{r}
safe <- iiss
iiss <- safe

# View NAs in equipment_type column
na_checks <- iiss %>% dplyr::filter(is.na(equipment_type))
sort(unique(na_checks$equipment_subtype))
sort(unique(na_checks$equipment_name))
sort(unique(na_checks$equipment_subname))

iiss <- iiss %>% dplyr::filter(!(is.na(country) & is.na(year) & is.na(equipment_type)))
```

## Old Row 
```{r}
iiss$unit_count[iiss$old_row == 1] <- 0

oldrow_checks <- iiss %>% 
  dplyr::filter(old_row == 1 &
                  (unit_count > 0))
```

# Post-Wrangling Fixes
## Unique unit types by equipment type 
```{r}
# Work in progress
abb <- as.data.frame(sort(unique(iiss$equipment_name))) %>%
  dplyr::rename(iiss_platform_abbrev = "sort(unique(iiss$equipment_name))")

abb_match <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "iiss_platforms_dedup")

abb_match <- abb_match %>%
  dplyr::filter(is.na(delete)) %>%
  dplyr::select(iiss_platform_abbrev, iiss_platform_name)

abb_match$original <- 1

abbs <- dplyr::left_join(abb, abb_match)

write.csv(abbs, paste0(here::here(), "/data/aux/","equipment_name_abbrevs.csv"))
```

## Non-alphanumeric values 
```{r}
# list of entries with non-alphanumeric values
n_alpha_service <- iiss %>%
  dplyr::filter(stringr::str_detect(service, '\\/|\\|\\-')) %>%
  dplyr::distinct(service)
iiss$service <- stringr::str_replace(iiss$service, "/", " and ")

n_alpha_subservice <- iiss %>%
  dplyr::filter(stringr::str_detect(subservice, '\\/|\\|\\-')) %>%
  dplyr::distinct(subservice)
iiss$subservice <- stringr::str_replace(iiss$subservice, "/", " and ")

n_alpha_subsubservice <- iiss %>%
  dplyr::filter(stringr::str_detect(subsubservice, '\\/|\\|\\-')) %>%
  dplyr::distinct(subsubservice)

n_alpha_equipment_type <- iiss %>%
  dplyr::filter(stringr::str_detect(equipment_type, '\\/|\\|\\-')) %>%
  dplyr::distinct(equipment_type)

n_alpha_equipment_subtype <- iiss %>%
  dplyr::filter(stringr::str_detect(equipment_subtype, '\\/|\\|\\-')) %>%
  dplyr::distinct(equipment_subtype)

n_alpha_equipment_name <- iiss %>%
  dplyr::filter(stringr::str_detect(equipment_name, '\\/|\\|\\-')) %>%
  dplyr::distinct(equipment_name)

n_alpha_equipment_subname <- iiss %>%
  dplyr::filter(stringr::str_detect(equipment_subname, '\\/|\\|\\-')) %>%
  dplyr::distinct(equipment_subname)

n_alpha_unit_name <- iiss %>%
  dplyr::filter(stringr::str_detect(unit_name, '\\/|\\|\\-')) %>%
  dplyr::distinct(unit_name)

n_alpha_total <- qpcR:::cbind.na(n_alpha_service, n_alpha_subservice, n_alpha_subsubservice,
                        n_alpha_equipment_type, n_alpha_equipment_subtype, n_alpha_equipment_name,
                        n_alpha_equipment_subname, n_alpha_unit_name)
write.csv(n_alpha_total)
```

## Column format coercion
```{r}
iiss$year <- as.numeric(iiss$year)
iiss$country <- as.character(iiss$country)
iiss$ccode <- as.character(iiss$ccode)
iiss$service <- as.character(iiss$service)
iiss$subservice <- as.character(iiss$subservice)
iiss$subsubservice <- as.character(iiss$subsubservice)
iiss$equipment_type <- as.character(iiss$equipment_type)
iiss$equipment_subtype <- as.character(iiss$equipment_subtype)
iiss$equipment_name <- as.character(iiss$equipment_name)
iiss$equipment_subname <- as.character(iiss$equipment_subname)
iiss$unit_name <- as.character(iiss$unit_name)
iiss$unit_count <- as.numeric(iiss$unit_count)
iiss$rawtext <- as.character(iiss$rawtext)
iiss$new_row <- as.numeric(iiss$new_row)
iiss$notes <- as.character(iiss$notes)
iiss$old_row <- as.numeric(iiss$old_row)
```

## Overlapping values between columns
```{r}
sort(unique(iiss$equipment_type))
sort(unique(iiss$equipment_subtype))
sort(unique(iiss$equipment_name))
sort(unique(iiss$equipment_subname))

# Identify unique values in each column
uet <- unique(iiss$equipment_type)
uest <- unique(iiss$equipment_subtype)
uen <- unique(iiss$equipment_name)
uesn <- unique(iiss$equipment_subname)

# Combine them into a single vector
uniques <- c(uest, uen)

unique_df <- as.data.frame(uniques) %>%
  dplyr::distinct() %>%
  dplyr::arrange(uniques)

# See which values are duplicated across the 4 columns
uniques <- sort(unique(uniques[duplicated(uniques)]))

uniques_df <- subset(iiss, equipment_type %in% c(uniques) |
                 equipment_subtype %in% c(uniques) |
                 equipment_name %in% c(uniques) |
                 equipment_subname %in% c(uniques)) %>% 
  dplyr::distinct()

as.data.frame(uniques) %>%
  DT::datatable()
```

# Country-year-unit df
## Create
Creates a dataframe for equipment unit counts per country/year.
```{r}
df <- country_year_equip_df(iiss)
df
saveRDS(df,paste0(here::here(),"/data/aux/","Country_Year.rds"))
```

# Save cleaned, final dataset
```{r}
saveRDS(iiss, paste0(here::here(), "/data/","01b_clean.rds"))

# Save a separate version without the unnecessary columns for the first version of the data
iiss_raw <- iiss %>%
  dplyr::select(-c(rawtext, new_row, old_row, notes)) %>%
  dplyr::filter(is.na(unit_count) | unit_count > 0) %>%
  dplyr::filter(year >= 1970 & year <= 2014)

saveRDS(iiss_raw, paste0(here::here(), "/data/", "rDMC_raw_v1.rds"))
write.csv(iiss_raw, paste0(here::here(), "/data/", "rDMC_raw_v1.csv"))
```

# System info
```{r}
sessionInfo()
```
