==> docs/04_prep_subjective.Rmd <==
---
title: "01d Prep Subjective"
author: "J Andres Gannon"
date: "5/1/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(magrittr)
library(stringr)
```

# Load
This is the data that is aggregated to the equipment_name level
```{r}
rm(list = ls())

df <- readRDS(paste0(here::here(), "/data/","01c_aggreg.rds"))

# New column to play with
df$tek_aggreg <- df$tek

# Basics
length(unique(df$tek))
sum(is.na(df$tek)) # NAs have been correctly converted to 0's
sort(unique(df$equipment_type))
```

# Manual aggregations
Organize manual aggregation changes by the equipment_type level

## air defence
The 4 types of air defence should be:
- surface to air missiles
- radar
- air to air missiles
- air to air artillery
```{r}
list <- sort(unique(str_subset(df$tek_aggreg, "air defence_")))
list <- as.data.frame(list)

# Move teks from other categories into here, given we are creating new subtypes
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_AA_TOWED" = "air defence_AA_TOWED"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_AA_SP" = "air defence_AA_SP"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_AD GUNS_SP" = "air defence_AD GUNS_SP"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_GUNS_AD" = "air defence_GUNS_AD"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_NA_AA SP" = "air defence_NA_AA SP"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_NA_AD GUNS" = "air defence_NA_AD GUNS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_anti-aircraft_AA" = "air defence_NA_AA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_NA_AA" = "air defence_NA_AA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_NA_AAM" = "air defence_NA_AAM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_NA_AD" = "air defence_NA_AD"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_GUNS_AA GUNS" = "air defence_GUNS_AA GUNS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_GUNS_AA" = "air defence_GUNS_AA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_missiles_SAM" = "air defence_missiles_SAM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_SP_AA" = "air defence_SP_AA"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air-launched missiles_missiles_AAM" = "air defence_missiles_AAM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air-launched missiles_NA_AAM" = "air defence_NA_AAM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air-launched missiles_NA_AD" = "air defence_NA_AD"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("anti-tank/anti-infrastructure_SAM_NA" = "air defence_missiles_SAM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("anti-tank/anti-infrastructure_NA_AA" = "air defence_NA_AA"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("surface-to-air-missiles_NA_SAM" = "air defence_missiles_SAM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("surface-to-air-missiles_missiles_SAM" = "air defence_missiles_SAM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("surface-to-surface missile launchers_missiles_MANPAD" = "air defence_missiles_MANPAD"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("surface-to-surface missile launchers_missiles_SAM" = "air defence_missiles_SAM"))

# Move teks from here into other categories
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_missiles_ASM" = "artillery_missiles_ASM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_NA_ASM" = "artillery_missiles_ASM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_missiles_ASSM" = "artillery_missiles_ASSM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_MOR_NA" = "artillery_MOR_NA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_RCL_NA" = "artillery_RCL_NA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_NA_ASM" = "artillery_NA_ASM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_SSM_NA" = "artillery_surface to surface missiles_SSM"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_NA_AShM	" = "land/sea defence_NA_AShM	"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radar_AA" = "radars_radar_AA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radar_AA artillery fire control" = "radars_radar_AA artillery fire control"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radar_height-finder" = "radars_radar_height-finder"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radar_missile control" = "radars_radar_missile control"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_NA_radar" = "radars_NA_radar"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_NA_RADAR" = "radars_NA_RADAR"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radar_early warning" = "radars_radar_early warning"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radar_NA" = "radars_radar_NA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radar_RADAR" = "radars_radar_RADAR"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radar_surveillance" = "radars_radar_surveillance"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_land-based systems_SAM RADARS" = "radars_land-based systems_SAM RADARS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_land-based systems_early warning" = "radars_land-based systems_early warning"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_land-based systems_fire control" = "radars_land-based systems_fire control"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_NA_ABM" = "radars_NA_ABM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_NA_early warning" = "radars_NA_early warning"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_NA_EW" = "radars_NA_EW"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air defence_radars_NA" = "radars_radars_NA"))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "air defence_")))
list <- as.data.frame(list)

# Categorization of similar types
air_defence <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "air defence")

air_defence <- air_defence[order(air_defence$tek), ]
air_defence$insheet <- 1

dim(list)
dim(air_defence)

mismatch <- dplyr::anti_join(list, air_defence, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, air_defence, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with air_defence_ + 'aircraft$new_type'
combo$tek_aggreg_new <- paste0("air defence_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Confirm final
sort(unique(str_subset(df$tek_aggreg, "air defence_")))
```

## aircraft
```{r}
# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "aircraft_")))
list <- as.data.frame(list)

# Categorization of similar types
aircraft <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "aircraft")

aircraft <- aircraft[order(aircraft$tek), ]
aircraft$insheet <- 1

dim(list)
dim(aircraft)

mismatch <- dplyr::anti_join(list, aircraft, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, aircraft, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# If df$tek_aggreg has 'aircraft_NA' replace that string with aircraft_ + 'aircraft$new_type'
combo$tek_aggreg_new <- combo$tek_aggreg
combo$tek_aggreg_new <- str_replace(combo$tek_aggreg_new, "NA_", paste0(combo$new_type, "_"))

# Subset to new second level
combo$tek_aggreg_new <- sub("_[^_]+$", "", combo$tek_aggreg_new)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "aircraft_")))
```

## amphibious
```{r}
# Move from other categories into here
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_AKSH" = "amphibious_NA_AKSH"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_LCC" = "amphibious_landing ships_LCC"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_LCM" = "amphibious_landing ships_LCM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_LCT" = "amphibious_landing ships_LCT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_LCU" = "amphibious_landing ships_LCU"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_LPD" = "amphibious_landing ships_LPD"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_LSM" = "amphibious_landing ships_LSM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_LST" = "amphibious_landing ships_LST"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_SES PCI (trials)" = "amphibious_landing ship_SES"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_LCT" = "amphibious_NA_LCT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_LCU" = "amphibious_NA_LCU"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_LCVP" = "amphibious_NA_LCVP"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_LSM" = "amphibious_NA_LSM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_LST" = "amphibious_NA_LST"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_LS" = "amphibious_NA_LS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_landing_LCU" = "amphibious_landing_LCU"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_riverine_LC" = "amphibious_riverine_LC"))

# Move from here to other categories
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_AFDL" = "logistics and support_NA_AFDL"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_AFS" = "logistics and support_NA_AFS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_AGE" = "logistics and support_NA_AGE"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_AGOR" = "logistics and support_NA_AGOR"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_AK" = "logistics and support_NA_AK"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_APB" = "logistics and support_NA_APB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_ARD" = "logistics and support_NA_ARD"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_AXS" = "logistics and support_NA_AXS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_AX" = "logistics and support_NA_AX"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_DDS" = "logistics and support_NA_DDS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_LAK" = "logistics and support_NA_LAK"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_LOG" = "logistics and support_NA_LOG"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_SSA" = "logistics and support_NA_SSA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_SSAN" = "logistics and support_NA_SSAN"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_YDT" = "logistics and support_NA_YDT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_YFRT" = "logistics and support_NA_YFRT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_YTB" = "logistics and support_NA_YTB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_NA_YTT" = "logistics and support_NA_YTT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_dock_NA" = "logistics and support_dock_NA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("amphibious_special forces_DDS" = "logistics and support_special forces_DDS"))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "amphibious_")))
list <- as.data.frame(list)

# Categorization of similar types
amphib <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "amphibious")

amphib <- amphib[order(amphib$tek), ]
amphib$insheet <- 1

dim(list)
dim(amphib)

mismatch <- dplyr::anti_join(list, amphib, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, amphib, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with amphibious_ + 'aircraft$new_type'
combo$tek_aggreg_new <- paste0("amphibious_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Confirm final
sort(unique(str_subset(df$tek_aggreg, "amphibious_")))
```

## anti-tank/anti-infrastructure
```{r}
# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "anti-tank/anti-infrastructure_")))
list <- as.data.frame(list)

# Categorization of similar types
atai <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "atai")

atai <- atai[order(atai$tek), ]
atai$insheet <- 1

dim(list)
dim(atai)

mismatch <- dplyr::anti_join(list, atai, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, atai, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with atai_ + 'atait$new_type'
combo$tek_aggreg_new <- paste0("anti-tank/anti-infrastructure_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "anti-tank/anti-infrastructure_")))
```

## armoured fighting vehicles
```{r}
# Check
sort(unique(str_subset(df$tek_aggreg, "armoured fighting vehicles")))

# Move teks from other categories into here, given we are creating new subtypes
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_GUNS_ARV" = "armoured fighting vehicles_transport_ARV"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_NA_APC(T)" = "armoured fighting vehicles_NA_APC(T)"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_NA_APC" = "armoured fighting vehicles_NA_APC"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("anti-tank/anti-infrastructure_NA_RECCE" = "armoured fighting vehicles_NA_RECCE"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("anti-tank/anti-infrastructure_RCL_RECCE	" = "armoured fighting vehicles_RCL_RECCE"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("anti-tank/anti-infrastructure_RL_RECCE	" = "armoured fighting vehicles_RL_RECCE"))

df$tek_aggreg[agrep("engineering and maintenance vehicles_", df$tek_aggreg)] <- "engineering and maintenance vehicles_all"
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("engineering and maintenance vehicles" = "armoured fighting vehicles_engineering and maintenance"))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "armoured fighting vehicles_")))
list <- as.data.frame(list)

# Categorization of similar types
afv <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "afv")

afv <- afv[order(afv$tek), ]
afv$insheet <- 1

dim(list)
dim(afv)

mismatch <- dplyr::anti_join(list, afv, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, afv, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with afv + 'afv$new_type'
combo$tek_aggreg_new <- paste0("armoured fighting vehicles_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Confirm
sort(unique(str_subset(df$tek_aggreg, "armoured fighting vehicles")))
```

## ballistic missiles
```{r}
# Check
sort(unique(str_subset(df$tek_aggreg, "ballistic missiles")))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "ballistic missiles_")))
list <- as.data.frame(list)

# Categorization of similar types
ballistic <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "ballistic")

ballistic <- ballistic[order(ballistic$tek), ]
ballistic$insheet <- 1

dim(list)
dim(ballistic)

mismatch <- dplyr::anti_join(list, ballistic, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, ballistic, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with ballistic + 'ballistic$new_type'
combo$tek_aggreg_new <- paste0("ballistic missiles_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Confirm
sort(unique(str_subset(df$tek_aggreg, "ballistic missiles")))
```

## helicopters
Do the same thing I did with aircraft
```{r}
# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "helicopters_")))
list <- as.data.frame(list)

# Temp used to make google doc
# write.csv(list, paste0(here::here(), "/data/","helicopters.csv"))

# Categorization of similar types
heli <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "helicopters")

heli <- heli[order(heli$tek), ]
heli$insheet <- 1

dim(list)
dim(heli)

mismatch <- dplyr::anti_join(list, heli, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, heli, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# If df$tek_aggreg has 'helicopter_NA' replace that string with helicopter_ + 'heli$new_type'
combo$tek_aggreg_new <- combo$tek_aggreg
combo$tek_aggreg_new <- str_replace(combo$tek_aggreg_new, "NA_", paste0(combo$new_type, "_"))

# Subset to new second level
combo$tek_aggreg_new <- sub("_[^_]+$", "", combo$tek_aggreg_new)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "helicopters_")))
```

## land/sea defense
The 3 types of land/sea defense should be:
- surface to surface missiles
- surface to surface artillery
- air to surface missiles

Fix former categories that are now coerced into land/sea defence
```{r}
# Fix former air-launched missiles
sort(unique(str_subset(df$tek_aggreg, "air-launched missiles_")))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air-launched missiles_" = "land/sea defence_"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("air-launched missiles_land-based systems_fire control" = "radars_land-based systems_fire control"))

# Fix former bombs
sort(unique(str_subset(df$tek_aggreg, "bombs_")))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("bombs_NA_conventional" = "land/sea defence_air to surface missiles"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("bombs_NA" = "land/sea defence_air to surface missiles"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("bombs_other_infrared/TV" = "land/sea defence_air to surface missiles"))

# Fix former surface to surface missiles
sort(unique(str_subset(df$tek_aggreg, "surface-to-surface missile launchers_")))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("surface-to-surface missile launchers_missiles" = "land/sea defence_surface to surface missiles"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("surface-to-surface missile launchers_SSM_SSM" = "land/sea defence_surface to surface missiles_SSM"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("surface-to-surface missile launchers_NA_AShM" = "land/sea defence_surface to surface missiles_AShM"))
```

Fix new category, where artillery is the base
```{r}
# Rename artillery to land/sea defence
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("artillery_" = "land/sea defence_"))
sort(unique(str_subset(df$tek_aggreg, "land/sea defence_")))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("land/sea defence_land-based systems_fire control" = "radars_land-based systems_fire control"))

list <- sort(unique(str_subset(df$tek_aggreg, "land/sea defence_")))
list <- as.data.frame(list)

# Categorization of similar types
landsea_defence <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "land/sea defence")

landsea_defence <- landsea_defence[order(landsea_defence$tek), ]
landsea_defence$insheet <- 1

dim(list)
dim(landsea_defence)

mismatch <- dplyr::anti_join(list, landsea_defence, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, landsea_defence, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with land/sea_defence_ + 'df$new_type'
combo$tek_aggreg_new <- paste0("land/sea defence_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Confirm
sort(unique(str_subset(df$tek_aggreg, "land/sea defence_")))
```

## logistics and support
```{r}
# Move from here to other categories
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_DMS" = "mine warfare_mine countermeasures"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MCMV" = "mine warfare_mine countermeasures"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MCM" = "mine warfare_mine countermeasures"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MCS" = "mine warfare_mine countermeasures"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MHC" = "mine warfare_mine countermeasures"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MLC" = "mine warfare_mine layers"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MSC" = "mine warfare_mine countermeasures"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MSL" = "mine warfare_mine countermeasures"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MSO" = "mine warfare_mine countermeasures"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_CL" = "principal surface combatants_cruisers_light"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_BB" = "principal surface combatants_battleships_BB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_CV" = "principal surface combatants_aircraft carriers_CV"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_MTB$" = "patrol and coastal combatants_NA_MTB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_PCO$" = "patrol and coastal combatants_NA_PCO"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_PCR$" = "patrol and coastal combatants_NA_PCR"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_PT$" = "patrol and coastal combatants_corvette_PT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_YP$" = "patrol and coastal combatants_NA_YP"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_attack craft_NA" = "patrol and coastal combatants_attack craft_NA"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_FFG$" = "principal surface combatants_frigates_NA"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_SSA$" = "submarines_NA_SSA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("logistics and support_NA_SSAN$" = "submarines_NA_SSAN"))

# Move from other categories to here
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_ABU" = "logistics and support_NA_ABU"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_ACV" = "logistics and support_NA_ACV"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AEF" = "logistics and support_NA_AEF"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AFS" = "logistics and support_NA_AFS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AG" = "logistics and support_NA_AG"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AGB" = "logistics and support_NA_AGB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_patrol craft_AGB" = "logistics and support_patrol craft_AGB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AGHS" = "logistics and support_NA_AGHS"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AOT" = "logistics and support_NA_AOT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AOR" = "logistics and support_NA_AOR"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AR" = "logistics and support_NA_AR"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_ATF" = "logistics and support_NA_ATF"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AT" = "logistics and support_NA_AT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AVT" = "logistics and support_NA_AVT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_AK" = "logistics and support_NA_AK"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_fleet replenishing ships" = "logistics and support_NA_fleet replenishing ships"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_icebreakers_NA" = "logistics and support_icebreakers_NA"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_repair" = "logistics and support_NA_repair"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_SAR" = "logistics and support_NA_SAR"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_Survey" = "logistics and support_NA_Survey"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_tankers" = "logistics and support_NA_tankers"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_trials/research" = "logistics and support_NA_trials/research"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_VTB" = "logistics and support_NA_VTB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_WTGB" = "logistics and support_NA_WTGB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_YDT" = "logistics and support_NA_YDT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_YTB" = "logistics and support_NA_YTB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_YDT" = "logistics and support_NA_YDT"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_YP" = "logistics and support_NA_YP"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("submarines_NA_VDT" = "logistics and support_NA_VDT"))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "logistics and support_")))
list <- as.data.frame(list)

# Categorization of similar types
logistics <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "logistics and support")

logistics <- logistics[order(logistics$tek), ]
logistics$insheet <- 1

dim(list)
dim(logistics)

mismatch <- dplyr::anti_join(list, logistics, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, logistics, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with land/sea_defence_ + 'df$new_type'
combo$tek_aggreg_new <- paste0("logistics and support_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "logistics and support_")))
```

## mine warfare
```{r}
# Move into here
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_MCMV" = "mine warfare_mine countermeasures"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_NA_MSC" = "mine warfare_mine countermeasures_MSC"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal mine countermeasures_MGB" = "mine warfare_mine countermeasures_MGB"))
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("patrol and coastal combatants_mine countermeasures_PC" = "mine warfare_mine countermeasures"))

# Move from here
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("mine warfare_aircraft carriers_MCO" = "mine warfare_mine countermeasures"))

df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("mine warfare_patrol craft_NA" = "patrol and coastal combatants_patrol craft_NA"))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "mine warfare_")))
list <- as.data.frame(list)

# Categorization of similar types
mine_warfare <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "mine warfare")

mine_warfare <- mine_warfare[order(mine_warfare$tek), ]
mine_warfare$insheet <- 1

dim(list)
dim(mine_warfare)

mismatch <- dplyr::anti_join(list, mine_warfare, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, mine_warfare, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with land/sea_defence_ + 'df$new_type'
combo$tek_aggreg_new <- paste0("mine warfare_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "mine warfare_")))
```

## patrol and coastal combatants
```{r}
# Check types
sort(unique(str_subset(df$tek_aggreg, "^patrol and coastal combatants_")))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "patrol and coastal combatants_")))
list <- as.data.frame(list)

## Subset to only teks with _NA_. Create new second level values for those that aren't matchable
patrol <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "patrol")

patrol <- patrol[order(patrol$tek), ]

patrol$insheet <- 1

dim(list)
dim(patrol)

mismatch <- dplyr::anti_join(list, patrol, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, patrol, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with land/sea_defence_ + 'df$new_type'
combo$tek_aggreg_new <- paste0("patrol and coastal combatants_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "patrol and coastal combatants_")))
```

## principal surface combatants
```{r}
# Check types
sort(unique(str_subset(df$tek_aggreg, "^principal surface combatants_")))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "principal surface combatants_")))
list <- as.data.frame(list)

## Subset to only teks with _NA_. Create new second level values for those that aren't matchable
surface <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "psc")

surface <- surface[order(surface$tek), ]

surface$insheet <- 1

dim(list)
dim(surface)

mismatch <- dplyr::anti_join(list, surface, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, surface, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with land/sea_defence_ + 'df$new_type'
combo$tek_aggreg_new <- paste0("principal surface combatants_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "principal surface combatants_")))
```

## radars
```{r}
# Check types
sort(unique(str_subset(df$tek_aggreg, "radars_")))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "radars_")))
list <- as.data.frame(list)

## Subset to only teks with _NA_. Create new second level values for those that aren't matchable
radars <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "radars")

radars <- radars[order(radars$tek), ]

radars$insheet <- 1

dim(list)
dim(radars)

mismatch <- dplyr::anti_join(list, radars, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, radars, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with land/sea_defence_ + 'df$new_type'
combo$tek_aggreg_new <- paste0("radars_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "radars_")))
```

## submarines
```{r}
sort(unique(str_subset(df$tek_aggreg, "submarines_")))

# Examine
list <- sort(unique(str_subset(df$tek_aggreg, "submarines_")))
list <- as.data.frame(list)

## Subset to only teks with _NA_. Create new second level values for those that aren't matchable
submarines <- googlesheets4::read_sheet(ss = "https://docs.google.com/spreadsheets/d/1yf-dGAiicIGNScouedvOCET4s34GGl3OelXrgOPWq8g/edit?usp=sharing",
                                    sheet = "subs")

submarines <- submarines[order(submarines$tek), ]

submarines$insheet <- 1

dim(list)
dim(submarines)

mismatch <- dplyr::anti_join(list, submarines, by = c("list" = "tek"))
mismatch
```

Assuming no mismatches, can now merge the new categorizations
```{r}
#Merge
combo <- dplyr::left_join(list, submarines, by = c("list" = "tek"))
combo <- combo %>% dplyr::rename(tek_aggreg = list)

# Replace the string with land/sea_defence_ + 'df$new_type'
combo$tek_aggreg_new <- paste0("submarines_", combo$new_type)

# Merge back into main df
combo <- combo %>% dplyr::select(c("tek_aggreg", "tek_aggreg_new"))
df <- dplyr::left_join(df, combo)

df <- df %>% dplyr::mutate(tek_aggreg = dplyr::coalesce(tek_aggreg_new, tek_aggreg))
df$tek_aggreg_new <- NULL

# Check
sort(unique(str_subset(df$tek_aggreg, "submarines_")))
```

## unmanned aerial vehicles
```{r}
sort(unique(str_subset(df$tek_aggreg, "unmanned aerial vehicles_")))

# Collapse to new categories
df$tek_aggreg <- str_replace_all(df$tek_aggreg, c("unmanned aerial vehicles_NA_AAM" = "unmanned aerial vehicles_recon-surveil",
                                                  "unmanned aerial vehicles_NA_ATK" = "unmanned aerial vehicles_combat",
                                                  "unmanned aerial vehicles_NA_CISR" = "unmanned aerial vehicles_combat",
                                                  "unmanned aerial vehicles_NA_EW" = "unmanned aerial vehicles_recon-surveil",
                                                  "unmanned aerial vehicles_NA_ISR" = "unmanned aerial vehicles_recon-surveil",
                                                  "unmanned aerial vehicles_NA_RECCE/ATK" = "unmanned aerial vehicles_combat",
                                                  "unmanned aerial vehicles_NA_RECCE" = "unmanned aerial vehicles_recon-surveil",
                                                  "unmanned aerial vehicles_NA_recon-surveil" = "unmanned aerial vehicles_recon-surveil",
                                                  "unmanned aerial vehicles_NA_TAC" = "unmanned aerial vehicles_combat",
                                                  "unmanned aerial vehicles_NA_TPT" = "unmanned aerial vehicles_transport",
                                                  "unmanned aerial vehicles_NA_TRIALS" = "unmanned aerial vehicles_recon-surveil",
                                                  "unmanned aerial vehicles_NA_TUAV" = "unmanned aerial vehicles_recon-surveil"))

# Check
sort(unique(str_subset(df$tek_aggreg, "unmanned aerial vehicles_")))
```

# Penultimate check
```{r}
sort(unique(df$tek_aggreg))
length(unique(df$tek_aggreg))
```

# Aggregate!
```{r}
df_agg <- df %>%
  dplyr::group_by(country, 
                  year,
                  tek_aggreg) %>% 
  dplyr::summarize(unit_count = sum(unit_count, na.rm = TRUE))

df_agg$unit_count[df_agg$unit_count == 0] <- NA
```

# Final check
```{r}
sort(unique(df_agg$tek_aggreg))
length(unique(df_agg$tek_aggreg))

# Rename so other code doesn't break
df_agg <- df_agg %>% dplyr::rename(tek = tek_aggreg)

# Plot count distrib
library(ggplot2)
df_agg %>%
  dplyr::select(country, year, unit_count) %>%
  dplyr::mutate(count_log = log(unit_count)) %>%
  ggplot(aes(x = count_log)) + 
  geom_density() + 
  theme_bw()
```

# Save
```{r}
saveRDS(df_agg, paste0(here::here(), "/data/","01d_subjective.rds"))
```

# System info
```{r}
sessionInfo()
```
