==> docs/05_prep_addrows.Rmd <==
---
title: "01c3Add All Technologies"
author: "J Andres Gannon"
date: "03/24/18"
output:
  html_notebook:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  html_document:
    toc: yes
editor_options:
  chunk_output_type: inline
---
<style>
    body .main-container {
        max-width: 100%;
    }
</style>


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This document pulls the cleaned long-form data from 01b and adds in all equipments that existed for each year, regardless of whether a given state owns those equipments.

# Set up
```{r}
library(magrittr)
library(tidyverse)
```

# Load raw IISS data 
```{r}
df <- readRDS(paste0(here::here(), "/data/","01d_subjective.rds"))
```

## Reformat to include all existing teks for each year
```{r}
# Create df that notes which are NA, since those need to stay NA
df_dummy <- df
df_dummy$unknown <- NA
df_dummy$unknown[is.na(df_dummy$unit_count)] <- 1
df_dummy$unit_count <- NULL

# Crate new rows
df_sum <- df %>% 
  dplyr::select(-country) %>% 
  dplyr::group_by(tek, year) %>% 
  dplyr::summarize(count = sum(unit_count)) %>% 
  dplyr::select(-count)
```

## Create true country-year list
This is a list of all the recognized country-years from 1970-2014
```{r}
states <- read.csv(paste0(here::here(), "/inst/extdata/","system2016.csv")) %>%
  dplyr::mutate(in_cow = 1) %>%
  dplyr::filter(year >= 1970) %>%
  dplyr::select(-version)
```

Merge stuff here
```{r}
cy <- data.frame(expand.grid(country = c(unique(as.character(df$country))),
                 year = 1970:2021)) %>% 
  dplyr::left_join(df_sum)

full <- dplyr::left_join(cy, df)

# Replace 0's with NA's based on dummy df
full <- dplyr::left_join(full, df_dummy)
full$unit_count[is.na(full$unit_count) & is.na(full$unknown)] <- 0
full$unknown <- NULL

# Filter out empty country names
full <- full %>%
  dplyr::filter(!is.na(country))

# Put cow country codes and numbers into this
full$stateabb <- countrycode::countrycode(full$country, "country.name", "cowc")
full$ccode <- countrycode::countrycode(full$country, "country.name", "cown")

# Manually fix the cow code matches that countrycode doesn't get right based on date
## German Federal Republic
full$stateabb[full$country == "german federal republic" & full$year < 1991] <- "GFR"
full$ccode[full$country == "german federal republic" & full$year < 1991] <- 260

## Yemen
full$stateabb[full$country == "yemen" & full$year < 1991] <- "YAR"
full$ccode[full$country == "yemen" & full$year < 1991] <- 678

## Serbia
full$stateabb[full$country == "serbia" & full$year >= 2006] <- "YUG"
full$ccode[full$country == "serbia" & full$year >= 2006] <- 345

# Continue
check_match <- full %>%
  dplyr::filter(year <= 2021) %>%
  dplyr::group_by(country, year, stateabb, ccode) %>%
  dplyr::summarise(equipment = sum(unit_count,
                                   na.rm = TRUE),
                   .groups = "keep") %>%
  dplyr::ungroup() %>%
  dplyr::left_join(., states) %>%
  dplyr::select(year, country, stateabb, ccode, in_cow, equipment) %>%
  tidyr::replace_na(list(in_cow = 0)) %>%
  dplyr::distinct()

# Prep check match and full for merge
check_match <- check_match %>%
  dplyr::filter(in_cow == 1) %>%
  dplyr::select(country, year, stateabb, ccode) %>%
  dplyr::mutate(keep = 1)

final <- dplyr::left_join(full, check_match) %>%
  dplyr::filter(!is.na(keep))

# Drop country-years that are all zeros
duds <- final %>%
  dplyr::group_by(country, year) %>%
  dplyr::summarise(equipment = sum(unit_count),
                   .groups = "keep") %>%
  dplyr::ungroup() %>%
  dplyr::mutate(nullentry = dplyr::if_else(equipment == 0, 1, 0),
                nullentry = dplyr::if_else(is.na(nullentry), 0, nullentry)) %>%
  dplyr::select(country, year, nullentry)

finalfinal <- dplyr::left_join(final, duds) %>%
  dplyr::filter(nullentry == 0) %>%
  dplyr::select(year, country, stateabb, ccode, tek, unit_count)

full <- finalfinal
```

# Save
```{r}
saveRDS(full, paste0(here::here(), "/data/","01e_addrows.rds"))

length(unique(full$tek))
```

# System info
```{r}
sessionInfo()
```
