==> docs/06_prep_wide.Rmd <==
---
title: "01d IISS Wide"
author: "J Andres Gannon"
date: "03/24/18"
output:
  html_notebook:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  html_document:
    toc: yes
editor_options:
  chunk_output_type: inline
---
<style>
    body .main-container {
        max-width: 100%;
    }
</style>


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This document pulls the cleaned long-form data from 01b and 01c and aggregates it to the equipment_name level. 

# Set up
```{r}
rm(list = ls())
library(magrittr)
library(tidyverse)
```

# Load aggregated IISS data 
```{r}
iiss <- readRDS(paste0(here::here(), "/data/","01e_addrows.rds"))
```

# Reshape data from long- to wide-form
```{r}
# Drop disaggregated tek name columns
iiss_wide <- iiss %>%
  tidyr::spread(tek, unit_count)
```

# Minor clean
Drop ordnances, columns with no values
```{r}
# create df of saved cols 
drop <- iiss_wide[, colSums(iiss_wide[,5:73], na.rm = T) != 0]

# save country
cntry <- iiss_wide[, 1:4]

# rejoin data (rows have not changed )
test <- dplyr::bind_cols(cntry, drop)

iiss_wide <- test

iiss_wide <- iiss_wide %>%
  dplyr::rename(year = year...1,
                country = country...2,
                stateabb = stateabb...3,
                ccode = ccode...4) %>%
  dplyr::select(-year...5,
                -country...6,
                -stateabb...7,
                -ccode...8)

# Also, this trash column
iiss_wide$`<NA>` <- NULL

colnames(iiss_wide)
```

# Save data
```{r}
saveRDS(iiss_wide, paste0(here::here(), "/data/", "01f_wide.rds"))
```

# System info
```{r}
sessionInfo()
```
