==> docs/07_prep_checks.Rmd <==
---
title: "01e IISS Checks"
author: "J Andres Gannon"
date: "03/24/18"
output:
  html_document:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---
<style>
    body .main-container {
        max-width: 100%;
    }
</style>


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This documents checks for errors in the data

# Set up
```{r}
library(magrittr)
library(ggplot2)
```

# Load and prep IISS data
## Load long and wide
```{r}
df_long <- readRDS(paste0(here::here(), "/data/","01e_addrows.rds"))
df_wide <- readRDS(paste0(here::here(), "/data/","01f_wide.rds"))
```

# Subset
At this point, the data is only complete to 2014, so we take out the later years to avoid incorrect inferences
```{r}
df_long <- df_long %>%
  dplyr::filter(year < 2015)

df_wide <- df_wide %>%
  dplyr::filter(year < 2015)
```

# Column checks
## Country
```{r}
sort(unique(df_long$country))
length(unique(df_long$country))
```

## Year
```{r}
sort(unique(df_long$year))

df_long %>%
  dplyr::count(year)
```

## Teks
```{r}
sort(unique(df_long$tek))
length(unique(df_long$tek))
```

## Count
```{r}
df_long %>%
    dplyr::select(country, year, unit_count) %>%
    ggplot(aes(x = unit_count)) + 
    geom_density() + 
    labs(title = "Count of Capabilities", x = "Count", y = "Percent of Rows") +
    theme_bw()

df_long %>%
    dplyr::select(country, year, unit_count) %>%
    ggplot(aes(x = unit_count)) + 
    geom_density() + 
    scale_x_log10() +
    labs(title = "Count of Capabilities (log)", x = "Count (log)", y = "Percent of Rows") +
    theme_bw()
```

Look at teks with odd counts
```{r}
subset(df_long, tek == "principal surface combatants_conventional carriers" & unit_count > 0) %>%
  DT::datatable()
```

# Big picture
## Data Explorer
```{r, fig.width = 16, fig.height = 16}
DataExplorer::plot_intro(df_long)
```

## dataMaid
```{r, fig.width = 16, fig.height = 16}
dataMaid::check(df_long)
dataMaid::summarize(df_long)
```

## dlookr
```{r, fig.width = 16, fig.height = 16, warning=FALSE, error=FALSE}
DT::datatable(dlookr::diagnose_outlier(df_wide))
```

## ExPanDaR
```{r, fig.width = 16, fig.height = 16}
ExPanDaR::prepare_missing_values_graph(df_wide, ts_id = "year")
ExPanDaR::prepare_trend_graph(df_long, "year")
ExPanDaR::prepare_scatter_plot(df_long, x = "year", y = "unit_count")
```

## funModeling
```{r, fig.width = 16, fig.height = 16}
summary(funModeling::data_integrity(df_long))
DT::datatable(funModeling::df_status(df_long))
funModeling::plot_num(df_wide)
```

## naniar
```{r, fig.width = 16, fig.height = 8}
naniar::gg_miss_var(df_long)
```

## smartEDA
```{r, fig.width = 16, fig.height = 16}
DT::datatable(SmartEDA::ExpNumStat(df_wide))
SmartEDA::ExpNumViz(df_long)
```

## summarytools
```{r, fig.width = 16, fig.height = 16, plain.ascii = FALSE}
print(summarytools::dfSummary(df_wide), max.tbl.height = 300, method = "render")
summarytools::descr(df_wide, style = 'rmarkdown')
```

## visdat
```{r, fig.width = 16, fig.height = 16}
visdat::vis_miss(df_wide, sort_miss = TRUE)
visdat::vis_dat(df_wide)
```

# Save final
These are the final versions of the data that should be used for all summary stats, visualizations, and modeling
```{r}
saveRDS(df_long, paste0(here::here(), "/data/", "rDMC_long_v1.rds"))
saveRDS(df_wide, paste0(here::here(), "/data/", "rDMC_wide_v1.rds"))

write.csv(df_long, paste0(here::here(), "/data/", "rDMC_long_v1.csv"))
write.csv(df_wide, paste0(here::here(), "/data/", "rDMC_wide_v1.csv"))
```

# System info
```{r}
sessionInfo()
```
